core: split out non-MLIR-specific parser functionality to GenericParser
