+++ xdsl/parser/__init__.py
@@ -4,4 +4,6 @@
 # We need to skip it here to allow importing from here instead.
 from .affine_parser import *  # noqa: TID251
 from .attribute_parser import *  # noqa: TID251
+from .base_parser import *  # noqa: TID251
 from .core import *  # noqa: TID251
+from .generic_parser import *  # noqa: TID251
