+++ xdsl/parser/core.py
@@ -21,8 +21,8 @@ from xdsl.utils.exceptions import MultipleSpansParseError
 from xdsl.utils.lexer import Input, Span
 from xdsl.utils.mlir_lexer import MLIRLexer, MLIRTokenKind
 
-from .attribute_parser import AttrParser, Position  # noqa: TID251
-from .base_parser import ParserState  # noqa: TID251
+from .attribute_parser import AttrParser  # noqa: TID251
+from .generic_parser import ParserState, Position  # noqa: TID251
 
 
 @dataclass(eq=False)
